Initial de-crust
This brings everything up-to-date except for gradle which is held back at 8.14.3 because 9.0 removes VersionNumber. I'm choosing to break migration here because in later commits, all of the tests will be removed and this project will be made specific to DNSNet. I don't have the time to figure out how to test this project against every ancient version of gradle and AGP that Mozilla supported.